Fix non-Qt implementation of UByteArray::right
UEFITool uses Qt but UEFIExtract does not so this should alter only the latter (and any of the other non-Qt products in the UEFITool project)
This fixes "GUID store" (checksum in the report and the contents of the extracted .bin file).
This may alter tail value validation and contents for revision 1 volumes.
This may alter crc validation for "Gaid store" and "Fsys store".