Fix non-Qt implementation of UByteArray::toHex
UEFITool uses Qt but UEFIExtract does not so this should alter only the latter (and any of the other non-Qt products in the UEFITool project)
Hex digits should be [a-f] instead of [k-p].
The string result should not be reversed.
This fixes the appearance of hash values used mostly in the CPD region.