Don't skip intersecting compressed CPD partition
A compressed partition can have meta data to indicate the compressed length. Many partitions may be missing this info. They have a partition length that actually represents the uncompressed length. This uncompressed length may overlap the next partition. In that case, the partition was skipped. Instead it should be assumed that the next partition begins after the compressed data and therefore the current partition's length should be adjusted to fit. The new length may include some padding bytes (usually 0xFF) but the only way around that is to decompress the partition to get the actual compressed length.